requirements.txt
altair
numpy
pandas
pydeck
streamlit
umap-learn
pandas
plotly
scikit-learn
translators
matplotlib